requirements.txt
python-telegram-bot
setuptools~=49.6.0
lxml~=4.6.3